#9874: Overload binary sub in unary sub
